move poolFetch function to actions
